requirements.txt
PyQt6>=6.7.0
python-dateutil>=2.9.0